+++ src/prefabs/structures/CardMedia/index.ts
@@ -1,21 +1,22 @@
-import {
-  component,
-  PrefabComponentOption,
-  PrefabComponent,
-} from '@betty-blocks/component-sdk';
-import { options as defaults } from './options';
-
-type OptionProducer = (key: string) => PrefabComponentOption;
-
-export interface Configuration {
-  label?: string;
-  options?: Record<string, OptionProducer>;
-}
+import { component, PrefabReference } from '@betty-blocks/component-sdk';
+import { Configuration } from '../Configuration';
+import { cardMediaOptions, categories as defaultCategories } from './options';
 
 export const CardMedia = (
   config: Configuration,
-  descendants: PrefabComponent[] = [],
+  descendants: PrefabReference[] = [],
 ) => {
-  const options = { ...(config.options || defaults) };
-  return component('CardMedia', { options }, descendants);
+  const options = { ...(config.options || cardMediaOptions) };
+  const style = { ...config.style };
+  const ref = config.ref ? { ...config.ref } : undefined;
+  const label = config.label ? config.label : undefined;
+  const optionCategories = config.optionCategories
+    ? config.optionCategories
+    : defaultCategories;
+
+  return component(
+    'CardMedia',
+    { options, style, ref, label, optionCategories },
+    descendants,
+  );
 };