+++ src/prefabs/structures/SubmitButton/index.ts
@@ -1,24 +1,14 @@
-import {
-  component,
-  OptionProducer,
-  PrefabComponent,
-  PrefabComponentStyle,
-} from '@betty-blocks/component-sdk';
+import { component, PrefabReference } from '@betty-blocks/component-sdk';
+import { Configuration } from '../Configuration';
 import { options as defaults } from './options';
-
-export interface Configuration {
-  options?: Record<string, OptionProducer>;
-  style?: PrefabComponentStyle;
-  ref?: { id: string };
-}
 
 export const SubmitButton = (
   config: Configuration,
-  children: PrefabComponent[] = [],
+  descendants: PrefabReference[] = [],
 ) => {
   const options = { ...(config.options || defaults) };
   const style = { ...config.style };
   const ref = config.ref ? { ...config.ref } : undefined;
-
-  return component('Button', { options, style, ref }, children);
+  const label = config.label ? config.label : undefined;
+  return component('Button', { options, style, ref, label }, descendants);
 };