+++ src/prefabs/structures/index.ts
@@ -1,0 +1,244 @@
+// Import all prefabs here
+import { ActionButton } from './ActionButton';
+import { ActionJSButton } from './ActionJSButton';
+import { Alert, FormSuccessAlert, FormErrorAlert } from './Alert';
+import { AppBar } from './Appbar';
+import { AutocompleteInput } from './AutocompleteInput';
+import { Avatar } from './Avatar';
+import { Box } from './Box';
+import { BreadcrumbItem } from './BreadcrumbItem';
+import { Breadcrumbs } from './Breadcrumbs';
+import { Button } from './Button';
+import { Card } from './Card';
+import { CardActions } from './CardActions';
+import { CardContent } from './CardContent';
+import { CardHeader } from './CardHeader';
+import { CardMedia } from './CardMedia';
+import { Carousel } from './Carousel';
+import { CarouselImage } from './CarouselImage';
+import { Chip } from './Chip';
+import { Column } from './Column';
+import { Conditional } from './Conditional';
+import { Container } from './Container';
+import { DataContainer } from './DataContainer';
+import { DataList } from './DataList';
+import { DataTable } from './DataTable';
+import { DataTableColumn } from './DataTableColumn';
+import { DateTimePicker } from './DateTimePicker';
+import { DetailViewChild } from './DetailViewChild';
+import { Dialog } from './Dialog';
+import { Divider } from './Divider';
+import { Drawer } from './Drawer';
+import { DrawerBar } from './DrawerBar';
+import { DrawerContainer } from './DrawerContainer';
+import { ExpansionPanel } from './ExpansionPanel';
+import { Grid } from './Grid';
+import { Icon } from './Icon';
+import { List } from './List';
+import { ListItem } from './ListItem';
+import { ListSubHeader } from './ListSubHeader';
+import { Media } from './Media';
+import { Menu } from './Menu';
+import { MenuItem } from './MenuItem';
+import { MultiAutocomplete } from './MultiAutoCompleteInput';
+import { OpenPageButton } from './OpenPage';
+import { Panel } from './Panel';
+import { Paper } from './Paper';
+import { PriceInput } from './PriceInput';
+import { Progress } from './Progress';
+import { RadioInput } from './RadioInput';
+import { Row } from './Row';
+import { SelectInput } from './SelectInput';
+import { Snackbar } from './Snackbar';
+import { Step } from './Step';
+import { Stepper } from './Stepper';
+import { SubmitButton } from './SubmitButton';
+import { Tab } from './Tab';
+import { Tabs } from './Tabs';
+import { Text } from './Text';
+import { TextArea } from './TextArea';
+import { TextInput } from './TextInput';
+
+// Import all prefab options here
+import { options as actionButtonOptions } from './ActionButton/options';
+import { options as actionJSButtonOptions } from './ActionJSButton/options';
+import { options as alertOptions } from './Alert/options';
+import { options as appBarOptions } from './Appbar/options';
+import { options as autocompleteInputOptions } from './AutocompleteInput/options';
+import { options as avatarOptions } from './Avatar/options';
+import { options as boxOptions } from './Box/options';
+import { options as breadcrumbItemOptions } from './BreadcrumbItem/options';
+import { options as breadcrumbsOptions } from './Breadcrumbs/options';
+import { options as buttonOptions } from './Button/options';
+import { cardOptions } from './Card/options';
+import { cardActionsOptions } from './CardActions/options';
+import { cardContentOptions } from './CardContent/options';
+import { cardHeaderOptions } from './CardHeader/options';
+import { cardMediaOptions } from './CardMedia/options';
+import { options as carouselOptions } from './Carousel/options';
+import { options as carouselImageOptions } from './CarouselImage/options';
+import { options as chipOptions } from './Chip/options';
+import { options as columnOptions } from './Column/options';
+import { conditionalOptions } from './Conditional/options';
+import { options as containerOptions } from './Container/options';
+import { options as dataContainerOptions } from './DataContainer/options';
+import { options as dataListOptions } from './DataList/options';
+import { options as dataTableOptions } from './DataTable/options';
+import { options as dataTableColumnOptions } from './DataTableColumn/options';
+import { options as dateTimePickerOptions } from './DateTimePicker/options';
+import { options as detailViewChildOptions } from './DetailViewChild/options';
+import { options as dialogOptions } from './Dialog/options';
+import { options as dividerOptions } from './Divider/options';
+import { options as drawerOptions } from './Drawer/options';
+import { options as drawerBarOptions } from './DrawerBar/options';
+import { options as drawerContainerOptions } from './DrawerContainer/options';
+import { options as expansionPanelOptions } from './ExpansionPanel/options';
+import { options as gridOptions } from './Grid/options';
+import { options as iconOptions } from './Icon/options';
+import { listOptions } from './List/options';
+import { listItemOptions } from './ListItem/options';
+import { listSubHeaderOptions } from './ListSubHeader/options';
+import { options as mediaOptions } from './Media/options';
+import { options as menuOptions } from './Menu/options';
+import { options as menuItemOptions } from './MenuItem/options';
+import { options as multiAutocompleteOptions } from './MultiAutoCompleteInput/options';
+import { options as openPageButtonOptions } from './OpenPage/options';
+import { options as panelOptions } from './Panel/options';
+import { options as paperOptions } from './Paper/options';
+import { options as priceInputOptions } from './PriceInput/options';
+import { options as progressOptions } from './Progress/options';
+import { options as radioInputOptions } from './RadioInput/options';
+import { options as rowOptions } from './Row/options';
+import { options as selectInputOptions } from './SelectInput/options';
+import { options as snackbarOptions } from './Snackbar/options';
+import { options as stepOptions } from './Step/options';
+import { options as stepperOptions } from './Stepper/options';
+import { options as submitButtonOptions } from './SubmitButton/options';
+import { options as tabOptions } from './Tab/options';
+import { options as tabsOptions } from './Tabs/options';
+import { options as textOptions } from './Text/options';
+import { options as textAreaOptions } from './TextArea/options';
+import { options as textInputOptions } from './TextInput/options';
+
+export {
+  ActionJSButton,
+  ActionButton,
+  Alert,
+  AppBar,
+  AutocompleteInput,
+  Avatar,
+  Box,
+  BreadcrumbItem,
+  Breadcrumbs,
+  Button,
+  Card,
+  CardActions,
+  CardContent,
+  CardHeader,
+  CardMedia,
+  Carousel,
+  CarouselImage,
+  Chip,
+  Column,
+  Conditional,
+  Container,
+  DataContainer,
+  DataList,
+  DataTable,
+  DataTableColumn,
+  DateTimePicker,
+  DetailViewChild,
+  Dialog,
+  Divider,
+  Drawer,
+  DrawerBar,
+  DrawerContainer,
+  ExpansionPanel,
+  FormErrorAlert,
+  FormSuccessAlert,
+  Grid,
+  Icon,
+  List,
+  ListItem,
+  ListSubHeader,
+  Media,
+  Menu,
+  MenuItem,
+  MultiAutocomplete,
+  OpenPageButton,
+  Panel,
+  Paper,
+  PriceInput,
+  Progress,
+  RadioInput,
+  Row,
+  SelectInput,
+  Snackbar,
+  Step,
+  Stepper,
+  SubmitButton,
+  Tab,
+  Tabs,
+  Text,
+  TextArea,
+  TextInput,
+  actionJSButtonOptions,
+  actionButtonOptions,
+  alertOptions,
+  appBarOptions,
+  autocompleteInputOptions,
+  avatarOptions,
+  boxOptions,
+  breadcrumbItemOptions,
+  breadcrumbsOptions,
+  buttonOptions,
+  cardOptions,
+  cardActionsOptions,
+  cardContentOptions,
+  cardHeaderOptions,
+  cardMediaOptions,
+  carouselOptions,
+  carouselImageOptions,
+  chipOptions,
+  columnOptions,
+  conditionalOptions,
+  containerOptions,
+  dataContainerOptions,
+  dataListOptions,
+  dataTableOptions,
+  dataTableColumnOptions,
+  dateTimePickerOptions,
+  detailViewChildOptions,
+  dialogOptions,
+  dividerOptions,
+  drawerOptions,
+  drawerBarOptions,
+  drawerContainerOptions,
+  expansionPanelOptions,
+  gridOptions,
+  iconOptions,
+  listOptions,
+  listItemOptions,
+  listSubHeaderOptions,
+  mediaOptions,
+  menuOptions,
+  menuItemOptions,
+  multiAutocompleteOptions,
+  openPageButtonOptions,
+  panelOptions,
+  paperOptions,
+  priceInputOptions,
+  progressOptions,
+  radioInputOptions,
+  rowOptions,
+  selectInputOptions,
+  snackbarOptions,
+  stepOptions,
+  stepperOptions,
+  submitButtonOptions,
+  tabOptions,
+  tabsOptions,
+  textOptions,
+  textAreaOptions,
+  textInputOptions,
+};